add condition to destroy relationship if the current_user is the follower in the relationship

+++ app/controllers/relationships_controller.rb
@@ -6,7 +6,7 @@ class RelationshipsController < ApplicationController
 
   def destroy
     relationship = Relationship.find(params[:id])
-    relationship.destroy
+    relationship.destroy if current_user == relationship.follower
     redirect_to people_path
   end
 end
